Hoist SearchHelper out of the Freelex namespace, since it's used by Signbank and Freelex
